scoperegistry allows extra scopes to be added on the command line now:
$ scoperegistry some/path/Runtime.ini some/other/path/Fred.ini Joe.ini

This loads Fred and Joe scopes in addition to any scopes picked up
via the normal configuration. If Fred or Joe appear in configuration
as well as on the command line, the config file on the command line
takes precedence.

Minor clean-up for incorrectly named #include guards.

Changed name and location of config.h, and replaced the #define
with a const char* inside the scopes namespace.

Minor clean-up around cmake, to get better consistency.

Default location for Runtime.ini for the moment is
/usr/share/unity-scopes-api/Runtime.ini. The constant
is generated via the top-level CMakeLists.txt, so changing
the path there should change it everywhere.