Added @cond/@endcond for benchmarking classes to suppress doxygen complaints. We have a bug open for this, so we don't lose track of it:
https://bugs.launchpad.net/unity-scopes-api/+bug/1285932

Added headers in testing to the headers fed up towards the root,
so the doc target gets a dependency on these headers and rebuilds
if one of them changes.

I left the tutorial text alone, so we still get warnings for the places
where the totorial hasn't been updated for the new names.

Small fix to a doxygen comment.

Removed two old macros from CMakeLists.txt files that are no longer needed.