Changed StopPublisher to be resilient if the zmq context is terminated before the publisher is stopped. Added ability to wait for publisher to stop.
Changed ScopeProcess to use a weak pointer to publisher. By
the time the RegistryObject goes out of scope, the middleware and the
context are destroyed already, so there is no point in having the entries
in the process table hold the publisher alive.

Changed ObjectAdapter to always terminate the stop publisher correctly.

Made Registry test depend on scoperunner target, so we can't run the test
with an out-of-date scoperunner.