Fixes for bug #1393438: "Enable location data" string displays untranslated. Added internal.location.precision settings metadata with possible values None, Approximate, and Precise. This allows the shell to treat this setting as "special" and to supply the correct localized strings for the display name and the possible display values for the setting. I left the original internal.location setting in place, to allow for a smooth transition from the old hack (boolean), to the new hack (enum). Eventually, we need to back out the hacks altogether because, ideally, the shell should only pay attention to the LocationDataNeeded metadata attribute.
While fixing this, I found another bug: for setting schemas parsed from
an .ini file, the enumerator choices were in a dictionary entry called
"displayValues" but, when parsing from an equivalent JSON string, the
dictionary entry was called "values". "displayNames" is correct, so that's
what the code now does, regardless of whether the schema is written in
JSON or as a .ini file. The only affected part of the code is the
smartscopes proxy, which is the only part of the system that gets JSON
settings definitions.